refactor(frontend): decompose QuizTaker into taker-scoped modules
QuizTaker.tsx shrank from 531 lines to 185 lines. The component now
acts as a thin orchestrator that owns answer/submit state and delegates
data loading + rendering to focused modules under
`components/quiz/taker/`:

- types.ts          — `QuizAnswer` / `AnswerMap`
- useQuizTaker.ts   — hook that loads the quiz + attempts (handles the
                      `quizId` targeting and in-progress vs completed
                      attempts)
- QuizHeader.tsx    — title/description/points/passing/attempts strip
- QuestionPrompt.tsx — single question with MCQ / T-F / SA / essay
                      variants (delegates essay to EssayAnswer)
- EssayAnswer.tsx   — long-form textarea with min-words word counter
- ResultsView.tsx   — pass/fail card + per-question review (handles
                      manual "pending review" states)
- PreviousAttempts.tsx — attempts list rendered below the form
- index.ts          — barrel exports

No functional/behavioral changes. tsc, eslint, vitest (85/85 — including
the 8 QuizTaker tests), and `npm run build` all pass.

Made-with: Cursor
