+++ frontend/src/components/quiz/taker/index.ts
@@ -1,0 +1,7 @@
+export { EssayAnswer } from "./EssayAnswer"
+export { PreviousAttempts } from "./PreviousAttempts"
+export { QuestionPrompt } from "./QuestionPrompt"
+export { QuizHeader } from "./QuizHeader"
+export { ResultsView } from "./ResultsView"
+export { useQuizTaker } from "./useQuizTaker"
+export type { AnswerMap, QuizAnswer } from "./types"
