+++ frontend/src/components/quiz/taker/types.ts
@@ -1,0 +1,6 @@
+export interface QuizAnswer {
+  selected_option_id?: string
+  text_answer?: string
+}
+
+export type AnswerMap = Record<string, QuizAnswer>
